# src/main/resources/schema/gade-strava-swagger-server.json
{
	"swagger": "2.0",
	"info": {
		"title": "Strava API Wrapper",
		"description": "Wraps useful Strava API calls",
		"version": "0.0.1"
	},
	"host": "localhost",
	"schemes": [
		"http"
	],
	"basePath": "/api/wrapper",
	"produces": [
		"application/json"
	],
	"parameters": {
		"page": {
			"name": "page",
			"in": "query",
			"description": "Page number. Defaults to 1.",
			"type": "integer"
		},
		"perPage": {
			"name": "per_page",
			"in": "query",
			"description": "Number of items per page. Defaults to 30.",
			"type": "integer",
			"default": 30
		},
		"cacheAction": {
			"name": "cache_action",
			"in": "query",
			"description": "Cache Action",
			"type": "string",
			"enum": ["none", "update", "renew"],
			"default": "none"
		}
		
	},	"paths": {
		"/gade/activities": {
			"get": {
				"operationId": "getActivities",
				"summary": "List Activities",
				"description": "Returns the activities",
				"parameters": [
				  {
					"$ref": "#/parameters/cacheAction"
				  },
				  {
				    "name": "after",
				    "in": "query",
				    "description": "An epoch timestamp to use for filtering activities that have taken place after a certain time.",
				    "type": "integer"
				  },
				  {
				    "$ref": "#/parameters/page"
				  },
				  {
				    "$ref": "#/parameters/perPage"
				  }
				],

				"tags": [
					"Activities"
				],
				"responses": {
					"200": {
						"description": "The authenticated athlete's activities",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "sub-schema/gade-activity.json#/GadeSummaryActivity"
							}
						}
					},
					"default": {
						"description": "Unexpected error.",
						"schema": {
							"$ref": "sub-schema/gade-fault.json#/GadeFault"
						}
					}
				}
			}
		},
		"/gear/{id}": {
			"get": {
				"operationId": "getGearById",
				"summary": "Get Equipment",
				"description": "Returns an equipment using its identifier.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "The identifier of the gear.",
						"required": true,
						"type": "string"
					}
				],
				"tags": [
					"Gears"
				],
				"responses": {
					"200": {
						"description": "A representation of the gear.",
						"schema": {
							"$ref": "sub-schema/gear.json#/DetailedGear"
						},
						"examples": {
							"application/json": {
								"id": "b1231",
								"primary": false,
								"resource_state": 3,
								"distance": 388206,
								"brand_name": "BMC",
								"model_name": "Teammachine",
								"frame_type": 3,
								"description": "My Bike."
							}
						}
					},
					"default": {
						"description": "Unexpected error.",
						"schema": {
							"$ref": "sub-schema/gade-fault.json#/GadeFault"
						}
					}
				}
			}
		}
	}
}